Expansion and Species Composition of an expansion domain

Expansion and Species Composition of an expansion domain

df_expansion(
  year,
  pool_f,
  species,
  bl,
  bl2,
  bl_unknown_method,
  bl_unknown_fished,
  bl_known_fished,
  p1,
  p2,
  sample_days,
  days,
  interviews,
  interviews_raw,
  catch,
  catch_raw,
  methods_refer,
  ports_refer,
  reference
)

Arguments

year
Year of the expansion

pool_f
Whether to pool interviews when fewer than 3 are available for a domain

species
a vector of species or all species (when spcies = NA)

bl
boat log entries ...

bl2
squared counts of boat log entries with known fishing method for each domain ([PORT_FK, TYPE_OF_DAY, METHOD_FK, CHARTER_F])

bl_unknown_method
counts of boat log entries with unknown fishing method for each domain ([PORT_FK, TYPE_OF_DAY, CHARTER_F])

bl_unknown_fished
counts of boat log entries with unknown fishing status for each domain ([PORT_FK, TYPE_OF_DAY, CHARTER_F])

bl_known_fished
counts of boat log entries with known fishing status for each domain ([PORT_FK, TYPE_OF_DAY, CHARTER_F])

p1
adjustment factors to account for missed vessels outside of shift times for each domain ([PORT_FK, METHOD_FK, CHARTER_F])

p2
adjustment factor to account for vessels outside of the sampled ports

sample_days
the number of sample days for each domain ([PORT_FK, TYPE_OF_DAY])

days
the number of calendar days for each domain ([TYPE_OF_DAY])

interviews
interviews for the expansion period

interviews_raw
all interviews for all years, for use when pooling

catch
interview catch data for the expansion period

catch_raw
interview catch data for all years

methods_refer
vector to convert method keys to method indices

ports_refer
vector to convert port keys to port indices

reference
reference values for use when insufficient interviews are available for a domain

Value

Two data frames containing all expansion and species composition fields for each expansion domain