Guam Survey Expansion Dataset

Total catch for each estimation domain. See the vignette "Guam Boat Based Expansion Database Fields" for more information.

g_bb_exp

Format

A data.frame with 8023 observations with 77 variables:

EXP_PK

The private key associated with a particular domain expansion.

PERIOD_TYPE

The time span of the expansion, with the only option being Y for annual

YEAR

The four digit year of the expansion

QUARTER

The quarter of the expansion, with the only option being 1 for an annual expansion.

TYPE_OF_DAY

The type of day of the expansion, either WD for weekday or WE for weekend/holiday.

PROCESS_TYPE

The source of interview data for the expansion, either G for actual interviews or R for reference values.

FISHERY_TYPE

The charter status of the expansion, either C for charter or N for non-charter.

PORT_FK

The database key corresponding to the port of the expansion

SURVEY_ZONE_FK

The database key corresponding to the survey zone of the expansion

METHOD_FK

The database key corresponding to the method key of the expansion

NUM_CALENDAR_DAYS

The total number of days within the expansion period.

NUM_SAMPLE_DAYS

The number of boat log sampling days within the expansion period.

NUM_INTERVIEW

The number of boat log entries (excluding METHOD_FK) where the fishing method was known.

NUM_INTERVIEW_POOLED

The number of boat log entries (excluding METHOD_FK) where the fishing method was unknown.

NUM_DAYS_FISHED

The total number of fishing days represented by interviews collected for the current domain.

NUM_TRIP

The number of boat log entries of a specfic fishing method for the current domain.

NUM_KN_METHOD_TRIP

The number of boat log entries (excluding METHOD_FK) where the fishing method was known.

NUM_UNKN_METHOD_TRIP

The number of boat log entries (excluding METHOD_FK) where the fishing method was unknown

NUM_KN_FISHED_TRIP

The number of boat log entries (excluding METHOD_FK) where it was known whether the trip was for fishing or not.

NUM_UNKN_FISHED_TRIP

The number of boat log entries (excluding METHOD_FK) where it was unknown whether the trip was for fishing or not.

SUM_EST_TRIP

The estimated total number of trips on sampling days for the current domain.

SUM_EST_TRIP2

The sum of the estimated square number of trips, aggregated by sample date.

SUM_HOUR

The total number of fishing hours represented on interviews for the current domain.

SUM_HOUR2

The sum of squares of the number of fishing hours represented on interviews for the current domain.

SUM_FISHER

The total number of fishers represented on interviews for the current domain.

SUM_FISHER2

The sum of squares of the number of fishers represented on interviews for the current domain.

SUM_FISHER_HOUR

The total number of fisher hours represented on interviews for the current domain.

SUM_FISHER_HOUR2

The sum of squares of the number of fisher hours represented on interviews for the current domain.

SUM_GEAR

The total number of fishing gears represented on interviews for the current domain.

SUM_GEAR2

The sum of squares of the number of fishing gears represented on interviews for the current domain.

SUM_TOT_KGS

The total catch represented on interviews for the current domain. Measured in kg.

SUM_TOT_KGS2

The sum of squares of the catch represented on interviews for the current domain. Measured in kg^2.

AVG_FISHER

The average number of fishers represented on interviews for the current domain.

AVG_HOUR

The average number of fishing hours represented on interviews for the current domain.

AVG_FISHER_HOUR

The average number of fisher hours represented on interviews for the current domain.

AVG_GEAR

The average number of fishing gears represented on interviews for the current domain.

AVG_EST_TRIP

The average estimated number of trips on sampling days for the current domain.

EXP_TRIP

The expanded number of trips for the current domain.

EXP_HOUR

The expanded number of fishing hours for the current domain.

EXP_FISHER

The expanded number of fishers for the current domain.

EXP_FISHER_HOUR

The expanded number of fisher-hours for the current domain. Measured in (fisher * hour).

EXP_GEAR

The expanded number of fishing gears for the current domain.

EXP_TOT_KGS

The expanded catch for the current domain.

CPUE_TRIP

The average catch-per-trip from interviews for the current domain.

CPUE_HOUR

The average catch-per-hour from interviews for the current domain. Measured in kg/hr.

CPUE_FISHER_HOUR

The average catch-per-fisher-hour from interviews for the current domain. Measured in kg/(fisher * hour).

CPUE_GEAR

The average catch-per-gear from interviews for the current domain. Measured in kg/gear.

VAR_AVG_FISHER

The variance of the number of fishers represented on interviews for the current domain.

VAR_AVG_HOUR

The variance of the number of fishing hours represented on interviews for the current domain.

VAR_AVG_FISHER_HOUR

The variance of the number of fisher hours represented on interviews for the current domain.

VAR_AVG_GEAR

The variance of the number of fishing gears represented on interviews for the current domain.

VAR_EST_TRIP

The variance of the estimated number of trips per sampling day for the current domain.

VAR_EXP_TRIP

The variance of the expanded number of trips for the current domain.

VAR_EXP_HOUR

The variance of the expanded number of fishing hours for the current domain.

VAR_EXP_FISHER

The variance of the expanded number of fishers for the current domain.

VAR_EXP_FISHER_HOUR

The variance of the expanded number of fisher-hours for the current domain. Measured in (fisher * hour)^2

VAR_EXP_GEAR

The variance of the expanded number of fishing gears for the current domain.

VAR_EXP_TOT_KGS

The variance of the expanded catch for the current domain. Measured in kg^2.

VAR_CPUE_TRIP

The variance of the catch-per-trip from interviews for the current domain. Measured in (kg/trip)^2.

VAR_CPUE_HOUR

The variance of the catch-per-hour from interviews for the current domain. Measured in (kg/hr)^2.

VAR_CPUE_GEAR

The variance of the catch-per-gear from interviews for the current domain. Measured in (kg/gear)^2.

VAR_CPUE_FISHER_HOUR

The variance of the catch-per-fisher-hour from interviews for the current domain. Measured in (kg/(fisher * hour))^2.

STD_EST_TRIP

The coefficient of variation of the average number of trips per sampling days for the current domain.

STD_EXP_TRIP

The coefficient of variation of the expanded number of trips for the current domain.

STD_EXP_HOUR

The coefficient of variation of the expanded number of fishing hours for the current domain.

STD_EXP_FISHER

The coefficient of variation of the expanded number of fishers for the current domain.

STD_EXP_FISHER_HOUR

The coefficient of variation of the expanded number of fisher-hours for the current domain. Measured in (fisher * hour).

STD_EXP_GEAR

The coefficient of variation of the expanded number of fishing gears for the current domain.

STD_EXP_TOT_KGS

The coefficient of variation of the expanded catch for the current domain. Measured in kg.

STD_CPUE_TRIP

The coefficient of variation of the catch-per-trip from interviews for the current domain. Measured in kg/trip.

STD_CPUE_HOUR

The coefficient of variation of the catch-per-hour from interviews for the current domain. Measured in kg/hr.

STD_CPUE_GEAR

The coefficient of variation of the catch-per-gear from interviews for the current domain. Measured in kg/gear.

STD_CPUE_FISHER_HOUR

The coefficient of variation of the catch-per-fisher-hour from interviews for the current domain. Measured in kg/(fisher * hour).

SYS_CREATE_TIME

System Time Entry was created.

SYS_MODIFY_TIME

System Time Entry was modified.

FLAG_POOLED

Flag to include pooled Interviews.

HASH

Hash signature.