Guam Species Composition Dataset

Representing the total catch by species for each estimation domain. See the vignette "Guam Boat Based Expansion Database Fields" for more information.

g_bb_spc

Format

A data.frame with 72049 observations with 25 variables:

SPC_PK

The private key associated with a particular species catch expansion.

EXP_FK

The private key associated with a particular domain expansion.

FLAG_POOLED

Flag to include pooled Interviews.

SPECIES_FK

The foreign key associated with a particular species.

NUM_INTERVIEW

The number of interviews of the current domain that included catch of the current species.

NUM_INTERVIEW_POOLED

The number of boat log entries (excluding METHOD_FK) where the fishing method was unknown.

KGS_CAUGHT

The estimated expanded catch for the current species s in the current domain.

NUM_KEPT

The number of specimens of the current species s that were recorded in interviews of the current domain.

PRICE_LB

Price per weight

SYS_CREATE_TIME

System Time Entry was created.

SYS_MODIFY_TIME

System Time Entry was modified.

HASH

Hash signature.

CPUE_TRIP

The average catch-per-trip from interviews for the current domain.

CPUE_HOUR

The average catch-per-hour from interviews for the current domain. Measured in kg/hr.

CPUE_GEAR

The average catch-per-gear from interviews for the current domain. Measured in kg/gear.

CPUE_FISHER_HOUR

The average catch-per-fisher-hour from interviews for the current domain. Measured in kg/(fisher * hour).

VAR_CPUE_TRIP

The variance of the catch-per-trip from interviews for the current domain. Measured in (kg/trip)^2.

VAR_KGS_CAUGHT

The variance of the estimated expanded catch for the current species s in the current domain.

STD_CPUE_TRIP

The coefficient of variation of the catch-per-trip from interviews for the current domain. Measured in kg/trip.

VAR_CPUE_HOUR

The variance of the catch-per-hour from interviews for the current domain. Measured in (kg/hr)^2.

STD_CPUE_HOUR

The coefficient of variation of the catch-per-hour from interviews for the current domain. Measured in kg/hr.

VAR_CPUE_GEAR

The variance of the catch-per-gear from interviews for the current domain. Measured in (kg/gear)^2.

STD_CPUE_HOUR.1

The coefficient of variation of the catch-per-hour from interviews for the current domain. Measured in kg/hr.

VAR_CPUE_FISHER_HOUR

The variance of the catch-per-fisher-hour from interviews for the current domain. Measured in (kg/(fisher * hour))^2.

STD_CPUE_FISHER_HOUR

The coefficient of variation of the catch-per-fisher-hour from interviews for the current domain. Measured in kg/(fisher * hour).